Add files via upload
This the project for Mobile Media Dev with Cindy Royal. My name is Jesse Williams. I changed sushi bonus to broccoli bonus. I did create an action that alerts the user and asks if they meant to select broccoli if they move the slider to 3. If they select yes that they mean to select brocoli the switch will move to on for them automatically. From there the slider adds a broccoli bonus instead of sushi bonus as noted in the tutorial for the setup. There was an additional step of adding an exception so that if the user selects no the broccolibonus would equal 0. There are 7 options to pick from, I added celery and ribs to the list, and added an additional output image to have picky eater, some foods, most food, and adventurous eater. I could not get the keyboard to popup when doing the text entry field but still put the Enter button in place to resignFirstResponder() when done typing. 

The only thing I wanted to add and had trouble getting it to work was to create a reset button that would take all of the switches back to false and the slider back to zero. No matter what method I tried to implement from what I could find online and looking for help, there was always some error being spit back that it was incorrect.